+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.1
+
+- Fixed bug affecting simple strings that begin & end with a variable.
+- Added InvalidFormatException for mismatched curly braces.
+
 ## 1.0.0
 
 - Initial version.